refactor: 컴파운드 패턴 합치기
