+++ src/components/form-control/FormControl.tsx
@@ -1,14 +1,18 @@
+import { useState } from 'react'
+import classNames from 'classnames'
 import {
   FormControlContext,
   useFormControlContext,
 } from '@/context/FormControlContext'
+import { Icon } from '@/components/icon'
 import type {
   FormControlProps,
   LabelProps,
   FieldProps,
 } from '@/types/form.types'
-import classNames from 'classnames'
 import styles from './FormControl.module.scss'
+
+// 기본 form-control field
 
 function Label({ children, visible = true }: LabelProps) {
   const { id } = useFormControlContext()
@@ -67,10 +71,7 @@ function ErrorMessage() {
   )
 }
 
-export default function FormControl({
-  children,
-  ...contextValue
-}: FormControlProps) {
+function FormControl({ children, ...contextValue }: FormControlProps) {
   return (
     <FormControlContext.Provider value={contextValue}>
       <div className={styles.container}>{children}</div>
@@ -81,3 +82,53 @@ export default function FormControl({
 FormControl.Label = Label
 FormControl.Field = Field
 FormControl.ErrorMessage = ErrorMessage
+
+// password form-control field
+
+function ToggleEye({
+  isPasswordVisible,
+  setIsPasswordVisible,
+}: {
+  isPasswordVisible: boolean
+  setIsPasswordVisible: React.Dispatch<React.SetStateAction<boolean>>
+}) {
+  const iconType = isPasswordVisible ? 'show' : 'hide'
+  const iconLabel = `비밀번호 ${isPasswordVisible ? '표시' : '감춤'}`
+
+  const handleToggle = () => {
+    setIsPasswordVisible((prevState) => !prevState)
+  }
+
+  return (
+    <button className={styles.button} type="button" onClick={handleToggle}>
+      <Icon type={iconType} alt={iconLabel} title={iconLabel} />
+    </button>
+  )
+}
+
+function PasswordField({ autoComplete, ...props }: FieldProps) {
+  const [isPasswordVisible, setIsPasswordVisible] = useState(false)
+
+  return (
+    <FormControl.Field
+      {...props}
+      type={isPasswordVisible ? 'text' : 'password'}
+      autoComplete={autoComplete}
+    >
+      <ToggleEye
+        isPasswordVisible={isPasswordVisible}
+        setIsPasswordVisible={setIsPasswordVisible}
+      />
+    </FormControl.Field>
+  )
+}
+
+function PasswordFormControl({ children, ...props }: FormControlProps) {
+  return <FormControl {...props}>{children}</FormControl>
+}
+
+PasswordFormControl.Label = FormControl.Label
+PasswordFormControl.Field = PasswordField
+PasswordFormControl.ErrorMessage = FormControl.ErrorMessage
+
+export { FormControl, PasswordFormControl }
