+++ src/context/FormControlContext.ts
@@ -1,14 +1,14 @@
 import { createContext, useContext } from 'react'
 import type {
   FieldContextType,
-  PasswordInputContextType,
+  PasswordFieldContextType,
 } from '@/types/form.types'
 
 export const FormControlContext = createContext<FieldContextType | undefined>(
   undefined,
 )
-export const PasswordInputContext = createContext<
-  PasswordInputContextType | undefined
+export const PasswordFormControlContext = createContext<
+  PasswordFieldContextType | undefined
 >(undefined)
 
 export function useFormControlContext() {
@@ -19,8 +19,8 @@ export function useFormControlContext() {
   return context
 }
 
-export function usePasswordInputContext() {
-  const context = useContext(PasswordInputContext)
+export function usePasswordFormControlContext() {
+  const context = useContext(PasswordFormControlContext)
   if (!context) {
     throw new Error(
       'usePasswordInputContext는 PasswordInput 내부에서만 사용 가능합니다.',
